update rlbygroupapi.sh
Made a fix on the logic for the sharedfilepath.

If an admin specified a location, then no actual filepath was created. (Which as you can imagine made things pretty difficult.) Found by: @joshnovotny

Corrected a typo in my credits for @joshnovotny. Really grateful for the insight on getting this script to run properly.